requirements.txt
boto==2.49.0
boto3==1.11.9
botocore==1.14.9
chardet==3.0.4
certifi==2020.6.20
Django==3.1.12
dj-database-url==0.5.0
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-package-monitor==0.6.1
django-storages==1.9.1
django-taggit==1.3.0
djangorestframework==3.12.4
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
Jinja2==2.11.3
Pillow==8.2.0
pipreqs==0.4.10
psutil==5.7.0
psycopg2==2.8.5 
python-dateutil==2.8.1 
pytz==2020.1
regex==2020.6.8
requests==2.24
requests-html==0.10.0
requests-oauthlib==1.3.0
requirements-parser==0.1.0
s3transfer==0.3.2
six==1.12.0
sqlparse==0.3.1
urllib3==1.26.5
websockets==9.1
whitenoise==5.1.0 